Added schedule logic in Plant_Cards_Fragment
Logic is currently in the Schedule button OnClick method. This logic should be moved wherever the "Add Schedule" button would be.
The button opens a form for the user to create a schedule entry for the given plant, and adds that to the database.
STILL TO DO (When the logic is put in the correct position): Get the correct plant ID in the ScheduleEntry object (right now it's just "bob").